requirements.txt
fastapi
uvicorn
jinja2
python-multipart
python-telegram-bot==20.7
pandas
openpyxl
